% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcars.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{Motor Trend Car Road Tests}
\format{
A data frame with 32 observations on 11 (numeric) variables:
\describe{
\item{\code{mpg}}{Miles/(US) gallon.}
\item{\code{cyl}}{integer. Number of cylinders.}
\item{\code{disp}}{Displacement (cu.in.).}
\item{\code{hp}}{Gross horsepower.}
\item{\code{drat}}{Rear axle ratio.}
\item{\code{wt}}{Weight (1000 lbs).}
\item{\code{qsec}}{1/4 mile time.}
\item{\code{vs}}{Engine (0 = V-shaped, 1 = straight).}
\item{\code{am}}{Transmission (0 = automatic, 1 = manual).}
\item{\code{gear}}{Number of forward gears.}
\item{\code{carb}}{Number of carburetors.}

@source The data is from the \code{mtcars} dataset in the
\code{link{dplyr}{mtcars}} dataset.
}
\usage{
mtcars
}
\description{
The data was extracted from the 1974 Motor Trend US magazine,
and comprises fuel consumption and 10 aspects of automobile design and performance for 32 automobiles (1973–74 models).
}
\keyword{datasets}
