% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Regression Function.R
\name{slreg}
\alias{slreg}
\title{Interactive Simple Linear Regression}
\usage{
slreg(data)
}
\arguments{
\item{data}{an object of class dataframe.}
}
\value{
an interactive shiny application for choosing outcome and independent variables from the dataset inputted to run simple linear regression on.
}
\description{
\code{slreg} builds an interactive simple linear regression application for the inputted dataset.
}
\details{
This function builds a user-friendly simple linear regression shiny application for the inputted dataset and includes model summary, data summary,
distribution of variables, linear assumptions plots, fitted and residuals graphs and table values, and correlation matrix of the variables.
}
\examples{
\dontrun{slreg(swiss)}
}
