% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise function.R
\name{stepreg}
\alias{stepreg}
\title{Interactive Stepwise Regression Model}
\usage{
stepreg(data)
}
\arguments{
\item{data}{an object of class data frame.}
}
\value{
an interactive shiny application for choosing outcome variable from the data set inputted to run stepwise regression.
}
\description{
\code{stepreg}builds an interactive stepwise regression application for the inputted data set.
}
\details{
This function builds an interactive stepwise regression application for the inputted data set and includes model summary, distribution of variables,
linear assumptions plots, fitted and residuals graphs and table values, and correlation matrix of the variables.
}
\examples{
\dontrun{stepreg(Boston)}

}
